Rename config for consistency.
