+++ src/repositories/embed.js
@@ -9,10 +9,10 @@ import config from '../../config';
 import { transformResponse } from './helpers';
 
 const embedClient = OEmbetter();
-const cache = new Cache(config('embeds.cache'));
+const cache = new Cache(config('embed.cache'));
 
 // Configure whitelisted domains & custom mappings. <https://git.io/fjeVb>
-embedClient.whitelist(config('embeds.whitelist'));
+embedClient.whitelist(config('embed.whitelist'));
 embedClient.endpoints(embedClient.suggestedEndpoints);
 
 // For sites where we don't support OEmbed, try OpenGraph/Twitter metatags:
@@ -20,7 +20,7 @@ embedClient.addBefore(async (url, options, _, callback) => {
   const { hostname } = parse(url);
 
   // We only prefer to fetch metatags for some domains:
-  const metatagDomains = config('embeds.preferMetatags');
+  const metatagDomains = config('embed.preferMetatags');
   if (!metatagDomains.some(domain => embedClient.inDomain(domain, hostname))) {
     return callback(null);
   }
